Migrate fluid precompile CommandController to symfony commands
